Add store.select(…) and update changeset
