+++ packages/xstate-store/src/index.ts
@@ -1,5 +1,4 @@
 export { shallowEqual } from './shallowEqual';
 export { fromStore } from './fromStore';
 export { createStore, createStoreWithProducer } from './store';
-export { select } from './select';
 export * from './types';
